Add pytest for date formatting
